Prepare 32 bit support

+++ src/emulator/serialization.hpp
@@ -378,7 +378,7 @@ namespace utils
         {
             const auto size = this->read<uint64_t>();
             result.clear();
-            result.reserve(size);
+            result.reserve(static_cast<size_t>(size));
 
             for (uint64_t i = 0; i < size; ++i)
             {
@@ -447,7 +447,7 @@ namespace utils
             const auto size = this->read<uint64_t>();
 
             result.clear();
-            result.reserve(size);
+            result.reserve(static_cast<size_t>(size));
 
             for (uint64_t i = 0; i < size; ++i)
             {
